Use numba for ~3x speedup in radial.py . Version bump. Details below
To get the numba speedup required inlining Lentz's method into the
evaluation of the continued fraction. This is not pretty and makes it
easier for a potential error to creep in, by not modularizing the
code. I added a doctest to check that the two approaches give the same
values.

Of course numba needed to be added to the requirements and to the
documentation noting the reqs. Similarly the documentation needed to
be updated, mostly because numba's decorator confuses autodoc.

In the process I improved the documentation in radial.py, but there is
still room for further improvement.